CLN: split test_indexing.py
split to: test_ix, test_scalar, test_loc, test_iloc, test_partial
